// src/lib/api-types.d.ts
export interface components {
    schemas: {
// ...
        NotificationResponseDto: {
            /** @example 1a0978f7-c1ca-45f2-bf15-8ca7358d2dc6 */
            id: string;
            /** @example Odemeniz gecikmekte! */
            title: string;
            /** @example Kredi karti odemenizin son tarihi gecmis durumda. */
            message: string;
            /**
             * @example CRITICAL_RED
             * @enum {string}
             */
            urgency: "LOW_GREEN" | "MEDIUM_YELLOW" | "CRITICAL_RED";
            /** @example false */
            isRead: boolean;
            /** @example 6f3335e9-1221-4908-a6d5-6209ec2e0877 */
            referenceId?: Record<string, never>;
            /**
             * Format: date-time
             * @example 2026-04-15T08:00:00.000Z
             */
            createdAt: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}